Revert method for determining NetCDF version
I need to re-think this because because the files aren't available when
I create the inventory. I can create the inventory on the machine
before inserts are done, or I can also just fill in NA for values I'm
not sure about and then try to find them at insert time.