Fix otherEntity adding
So basically the registry hates the DataONE API and the
physical@distribution@online@url has to be an ecogrid URL which needs
to reference a docID which is a Metacat concept.

Unfortunately, this adds an annoying requirement that adding
otherEntity elements to a metadata doc requires knowledge of which
Metacat instance you want the docID for and how it's accessed.

Here I implement a hack that guesses what to do based upon the
originMemberNode set in sysmeta.

Because of this I have to skip two tests (for now). :(