Add a check_first param to some possibly slow funcs
A few functions in editing.R (I may need to add this type of
functionality elsewhere) are slow because they check that all PIDs
passed in as args exist on the MN before starting their work. This is a
nice feature to prevent bad things but can slow operations down. It's
also unnecessary if the user just created the objects and knows they
exist.