Rewrite resource map updating routines to fix various issues
This commit addresses the cause of #51

The previous version of `parse_resource_map` used a SPARQL
query to pull the triples out of the resource map being updated
but didn't actually parse the result in an RDF-aware manner.
This lead to &quot; characters ending up in various places in
the RDF/XML which causes tons of issues.

The new routine simply uses datapack::parseRDF instead and
just filters out cito:documents, cito:isDocumentedBy, dcterms:identifier, and the DataONE R Client statement.