Make publish_update use the metadata object's formatId
Instead of hardcoding the formatID of the new object to EML 2.1.1, we instead just use the same formatID as what was set on the object we're obsoleting. We also add a format_id argument to allow an override of the behavior.

Also adds an example EML 220 doc and a format ID helper